renamed link plugin to anchor plugin, updated routes and components
