+++ docs/client/components/pages/Anchor/CustomInlineToolbarEditor/toolbarStyles.css
