+++ docs/client/components/pages/Anchor/SimpleLinkPluginEditor/editorStyles.css
