+++ docs/client/components/pages/Anchor/SimpleLinkPluginEditor/index.js
@@ -1,23 +1,12 @@
 import React, { Component } from 'react';
 import Editor, { createEditorStateWithText } from 'draft-js-plugins-editor';
 import createInlineToolbarPlugin from 'draft-js-inline-toolbar-plugin';
-import createLinkPlugin from 'draft-js-link-plugin';
-import {
-  ItalicButton,
-  BoldButton,
-  UnderlineButton,
-} from 'draft-js-buttons';
+import createLinkPlugin from 'draft-js-anchor-plugin';
+import { ItalicButton, BoldButton, UnderlineButton } from 'draft-js-buttons';
 import editorStyles from './editorStyles.css';
-import buttonStyles from './buttonStyles.css';
-import toolbarStyles from './toolbarStyles.css';
-import linkStyles from './linkStyles.css';
 
-const linkPlugin = createLinkPlugin({
-  theme: linkStyles,
-  placeholder: 'http://…'
-});
+const linkPlugin = createLinkPlugin();
 const inlineToolbarPlugin = createInlineToolbarPlugin({
-  theme: { buttonStyles, toolbarStyles },
   structure: [
     BoldButton,
     ItalicButton,
@@ -29,16 +18,16 @@ const { InlineToolbar } = inlineToolbarPlugin;
 const plugins = [inlineToolbarPlugin, linkPlugin];
 const text = 'Try selecting a part of this text and click on the link button in the toolbar that appears …';
 
-export default class ThemedInlineToolbarEditor extends Component {
+export default class SimpleLinkPluginEditor extends Component {
   state = {
     editorState: createEditorStateWithText(text)
   };
 
   onChange = (editorState) =>
-      this.setState({ editorState });
+    this.setState({ editorState });
 
   focus = () =>
-      this.editor.focus();
+    this.editor.focus();
 
   render() {
     return (
@@ -53,4 +42,4 @@ export default class ThemedInlineToolbarEditor extends Component {
       </div>
     );
   }
-  }
+}
