+++ docs/webpack.config.base.js
@@ -7,7 +7,7 @@ module.exports = {
       'draft-js-plugins-editor': path.join(__dirname, '..', 'draft-js-plugins-editor', 'src'),
       'draft-js-hashtag-plugin': path.join(__dirname, '..', 'draft-js-hashtag-plugin', 'src'),
       'draft-js-linkify-plugin': path.join(__dirname, '..', 'draft-js-linkify-plugin', 'src'),
-      'draft-js-link-plugin': path.join(__dirname, '..', 'draft-js-link-plugin', 'src'),
+      'draft-js-anchor-plugin': path.join(__dirname, '..', 'draft-js-anchor-plugin', 'src'),
       'draft-js-mention-plugin': path.join(__dirname, '..', 'draft-js-mention-plugin', 'src'),
       'draft-js-sticker-plugin': path.join(__dirname, '..', 'draft-js-sticker-plugin', 'src'),
       'draft-js-undo-plugin': path.join(__dirname, '..', 'draft-js-undo-plugin', 'src'),
@@ -41,7 +41,7 @@ module.exports = {
           path.join(__dirname, '..', 'draft-js-plugins-editor', 'src'),
           path.join(__dirname, '..', 'draft-js-hashtag-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-linkify-plugin', 'src'),
-          path.join(__dirname, '..', 'draft-js-link-plugin', 'src'),
+          path.join(__dirname, '..', 'draft-js-anchor-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-mention-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-sticker-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-undo-plugin', 'src'),
@@ -65,7 +65,7 @@ module.exports = {
           path.join(__dirname, '..', 'draft-js-plugins-editor', 'src'),
           path.join(__dirname, '..', 'draft-js-hashtag-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-linkify-plugin', 'src'),
-          path.join(__dirname, '..', 'draft-js-link-plugin', 'src'),
+          path.join(__dirname, '..', 'draft-js-anchor-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-mention-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-sticker-plugin', 'src'),
           path.join(__dirname, '..', 'draft-js-undo-plugin', 'src'),
