+++ draft-js-anchor-plugin/README.md
@@ -7,7 +7,7 @@ This plugin allows you to add link entities via the [inline toolbar](https://www
 ## Usage
 
 ```js
-import createLinkPlugin from 'draft-js-link-plugin';
+import createLinkPlugin from 'draft-js-anchor-plugin';
 
 const linkPlugin = createLinkPlugin();
 ```
@@ -15,6 +15,6 @@ const linkPlugin = createLinkPlugin();
 ### Importing the default styles
 
 The plugin ships with a default styling available at this location in the installed package:
-`node_modules/draft-js-link-plugin/lib/plugin.css`.
+`draft-js-anchor-plugin/lib/plugin.css`.
 
 If you want to use the default styles, you can include this stylesheet. Otherwise you can supply your own styles via the `theme` config option.
