+++ draft-js-anchor-plugin/src/linkStrategy.js
