+++ draft-js-anchor-plugin/src/utils/EditorUtils.js
