+++ draft-js-anchor-plugin/src/utils/URLUtils.js
