+++ draft-js-image-plugin/CHANGELOG.md
@@ -6,3 +6,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## To Be Released
 
 ### Released the first working of DraftJS State Plugin
+
+## 2.0.0-rc5 - 2016-08-16
+### added `extraData` parameter to addImage modifier, so additional parameters can be passed.
